🪵 Deduplicate exec/logging functions to use myst-cli-utils
