+++ apps/cli/src/cli/utils.ts
@@ -2,8 +2,8 @@ import fs from 'fs';
 import chalk from 'chalk';
 import check from 'check-node-version';
 import type { Command } from 'commander';
+import { chalkLogger, LogLevel } from 'myst-cli-utils';
 import { CURVENOTE_YML } from '../config/types';
-import { chalkLogger, LogLevel } from '../logging';
 import { Session, getToken } from '../session';
 import type { ISession } from '../session/types';
 import { selectors } from '../store';
