+++ apps/cli/src/export/utils/index.ts
@@ -4,4 +4,3 @@ export * from './getImageSrc';
 export * from './writeImagesToFiles';
 export * from './makeBuildPaths';
 export * from './assertions';
-export { default as exec, makeExecutable } from './exec';
