+++ apps/cli/src/sync/clone.ts
@@ -2,11 +2,11 @@ import chalk from 'chalk';
 import fs from 'fs';
 import inquirer from 'inquirer';
 import { join } from 'path';
+import { LogLevel } from 'myst-cli-utils';
 import type { SiteConfig } from '@curvenote/blocks';
 import { loadConfigOrThrow, writeConfigs } from '../config';
 import type { ProjectConfig } from '../config/types';
 import { projectIdFromLink } from '../export';
-import { LogLevel } from '../logging';
 import type { Project } from '../models';
 import type { ISession } from '../session/types';
 import { selectors } from '../store';
