+++ apps/cli/src/sync/init.ts
@@ -2,11 +2,11 @@ import chalk from 'chalk';
 import fs from 'fs';
 import inquirer from 'inquirer';
 import { basename, join, resolve } from 'path';
+import { LogLevel } from 'myst-cli-utils';
 import type { SiteNavPage } from '@curvenote/blocks';
 import { writeConfigs } from '../config';
 import { CURVENOTE_YML } from '../config/types';
 import { docLinks, LOGO } from '../docs';
-import { LogLevel } from '../logging';
 import { MyUser } from '../models';
 import type { ISession } from '../session/types';
 import { selectors } from '../store';
