+++ apps/cli/src/utils/utils.ts
@@ -6,9 +6,9 @@ import chalk from 'chalk';
 import inquirer from 'inquirer';
 import path from 'path';
 import prettyHrtime from 'pretty-hrtime';
+import type { Logger } from 'myst-cli-utils';
 import type { JsonObject, VersionId } from '@curvenote/blocks';
 import { configFileExists, loadConfigOrThrow } from '../config';
-import type { Logger } from '../logging';
 import type { ISession } from '../session/types';
 import type { WarningKind } from '../store/build';
 import { warnings } from '../store/build';
