+++ apps/cli/src/web/files.ts
@@ -1,8 +1,8 @@
 import fs from 'fs';
 import path from 'path';
+import type { Logger } from 'myst-cli-utils';
 import { KnownCellOutputMimeTypes } from '@curvenote/blocks';
 import type { IFileObject, IFileObjectFactoryFn, Metadata } from '@curvenote/nbtx';
-import type { Logger } from '../logging';
 import { computeHash } from '../utils';
 
 enum FileExtension {
