+++ packages/myst-cli-utils/src/index.ts
@@ -1,5 +1,12 @@
-export type { Logger, ISession } from './types';
-export { LogLevel, chalkLogger, silentLogger, createGitLogger, createNpmLogger } from './logger';
+export type { Logger, LoggerDE, ISession } from './types';
+export {
+  LogLevel,
+  basicLogger,
+  chalkLogger,
+  silentLogger,
+  createGitLogger,
+  createNpmLogger,
+} from './logger';
 export { exec, makeExecutable } from './exec';
 export { clirun, tic } from './utils';
 export { Session, getSession } from './session';
